% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_sads.R
\name{aggregate_sads}
\alias{aggregate_sads}
\title{aggregate SAD files from DiViMe and (manually) annotated ELAN .eaf files}
\usage{
aggregate_sads(sadloc, elanloc, segment_dur = 60, ...)
}
\arguments{
\item{sadloc}{character, the location of the DiViMe SAD results}

\item{elanloc}{character, the location of the ELAN .eaf files}

\item{segment_dur}{duration for the length of a segment (default is \code{60} (seconds))}

\item{...}{additional parameters for elan files (which tiers/annotations to ignore), see \code{\link{collapse_tiers}}}
}
\value{
a \code{data.frame} with columns for each of the SADs and the duration of speech per segment
}
\description{
for validation purposes
}
