% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babble_ratio.R
\name{babble_ratio}
\alias{babble_ratio}
\title{canonical babbling ratio}
\usage{
babble_ratio(xfile)
}
\arguments{
\item{xfile}{character, file path to rttm generated by
\code{\link{divime_classify_vox}}}
}
\value{
a list with three items (\code{cns}: number of canonical utterances,
\code{ncs}: number of non-canonical utterances, \code{babble_ratio}: the
ratio)
}
\description{
proportion of canonical utterances (out of the sum of canonical and
non-canonical)
}
\details{
if the rttm is empty (i.e. contains no child vocalizations at all),
the returned list will contain only \code{NA}s. If there were some child
vocalizations but none of them were CNS or NCS, the counts in the output will
be 0 and only the ratio is returned as \code{NA}.
}
