% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_filename.R
\name{clean_filename}
\alias{clean_filename}
\title{replace special characters with underscore}
\usage{
clean_filename(x)
}
\arguments{
\item{x}{character, typically a file name}
}
\value{
character, the file name with special characters replaced
}
\description{
produce file names that can be handled by the divime tools
}
\examples{
filename <- "a file (final).wav"
clean_filename(filename)
}
