% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_rttm.R
\name{combine_rttm}
\alias{combine_rttm}
\title{merge rttm files}
\usage{
combine_rttm(rttm_files, split_dur, basename = NULL)
}
\arguments{
\item{rttm_files}{character, paths to rttm files to merge}

\item{split_dur}{numeric, the duration of the underlying split audio}

\item{basename}{character, the name of the file (unmerged)}
}
\value{
a data.frame that can be written as rttm file
}
\description{
merge rttm files
}
