% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{create confusion matrix}
\usage{
confusion_matrix(test = NULL, gold = NULL)
}
\arguments{
\item{test}{numeric 0/1 vector for the test file, indicating whether there was speech or not at a given timepoint}

\item{gold}{numeric 0/1 vector for the gold standard file, indicating whether there was speech or not at a given timepoint}
}
\value{
a 2x2 matrix, where row labels refer to the gold standard and column lables to the tested/evaluated file
}
\description{
create confusion matrix
}
\examples{
x1 <- system.file("noisemesSad_synthetic_speech.rttm", package = "avutils")
x2 <- system.file("synthetic_speech.eaf", package = "avutils")
edata <- evaluate_sad(test = x1, reference = x2, overlap_res = 100,
                      raw_results = TRUE, doplot = FALSE)
confusion_matrix(test = edata[, "test_speech"], gold = edata[, "ref_speech"])
}
