% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_audio.R
\name{convert_audio}
\alias{convert_audio}
\title{convert audio file format}
\usage{
convert_audio(filein, outformat = list(filetype = "wav", res = 44100,
  mono = TRUE), overwrite = FALSE, pathout = NULL,
  pathtoffmpeg = getOption("avutils_ffmpeg"))
}
\arguments{
\item{filein}{character, paths to input audio files}

\item{outformat}{a list with specifications of the desired output format, see details}

\item{overwrite}{logical, should output files be overwritten if they already exist}

\item{pathout}{character, path were output is written, by default the same as the source}

\item{pathtoffmpeg}{character, path to ffmpeg binary}
}
\value{
a \code{data.frame} with diagnostics
}
\description{
convert from mp3 and possibly other audio file formats to .wav
}
\details{
The \code{outformat=} argument currently provides three options, and by default a mono (\code{mono = TRUE}) .wav files is produced (\code{filetpye = "wav"}) with 44.1kHz sampling rate (\code{res = 441000}).

Files will be potentially renamed, i.e. the file extension of the new output file will change to '.wav'.

Note that if the input file has the same name and is in the same location as the output (e.g. if you convert x.wav with \code{pathout = NULL}), nothing will happen even if you select \code{overwrite = TRUE} because \code{ffmpeg} won't overwrite its own source.
}
