% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_audio.R
\name{copy_audio}
\alias{copy_audio}
\title{copy audio files with optional silence added at the end}
\usage{
copy_audio(from, to, appendsilence = NULL,
  pathtosox = getOption("avutils_sox"))
}
\arguments{
\item{from}{character, file path for the source}

\item{to}{character, file path to the target location}

\item{appendsilence}{numeric, number of seconds for added silence at the end of the file. \code{NULL} by default where the file is simply copied without any processing through \code{sox}.}

\item{pathtosox}{character, path to sox binary}
}
\value{
logical, indicating whether the file \code{to} exists, i.e. whether copying was successful.
}
\description{
copy audio files with optional silence added at the end
}
\details{
In case you want to add silence to the audio file, \code{sox} is required (see \code{\link{set_binaries}} for details). In addition, the function creates an additional temporary copy of the wav file, because \code{sox} also requires 'clean' file names.
}
