% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divime_app.R
\name{divime_app}
\alias{divime_app}
\title{interactive user interface for avutils}
\usage{
divime_app(launchinbrowser = TRUE)
}
\arguments{
\item{launchinbrowser}{logical, should the app start in your default browser (by default \code{TRUE})}
}
\value{
launches a \code{shiny} app
}
\description{
interactive user interface for avutils
}
\examples{
\dontrun{
divime_app()
}
}
\author{
Christof Neumann
}
