% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divime_collate_output.R
\name{divime_collate_output}
\alias{divime_collate_output}
\title{combine .rttm output from DiViMe tools}
\usage{
divime_collate_output(data_loc)
}
\arguments{
\item{data_loc}{character, the location for the data folder (which may or may not be the same as the one where the source audio files are stored)}
}
\value{
a list with 8 components (for three SADs the actual SAD and corresponding diarization, yuni results and vcm)
}
\description{
combine .rttm output from DiViMe tools
}
