% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divime_diarization.R
\name{divime_diarization}
\alias{divime_diarization}
\title{run the DiViMe diarization module diartk}
\usage{
divime_diarization(audio_loc, divime_loc, speech_annos = "noisemes",
  vmstart = TRUE, vmshutdown = TRUE, messages = TRUE,
  overwrite = FALSE)
}
\arguments{
\item{audio_loc}{character, path to the audio files}

\item{divime_loc}{character, path to the DiViMe directory with running VM}

\item{speech_annos}{character, what kind of speech detection is available, see details}

\item{vmstart}{logical, perform a check whether the VM is running and if not start it up (by default \code{TRUE}). Turning this off, will speed up the function a little bit, but requires that you are sure that the VM is indeed running in \code{divime_loc}.}

\item{vmshutdown}{logical, should the VM be shut down after the operations are done (by default \code{TRUE})}

\item{messages}{logical, should the file names of each processed file be printed}

\item{overwrite}{logical, should output files be overwritten if they already exist (default is \code{FALSE})}
}
\value{
a data.frame with the locations of the created rttm files and some diagnostics
}
\description{
run the DiViMe diarization module diartk
}
\details{
\code{speech_annos} needs to be one of the following: \code{"noisemes"}, \code{"opensmile"}, \code{"tocombo"} or \code{"custom"}. Currently, \code{"custom"} is not yet implemented.
}
