% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divime_fullnoisemes.R
\name{divime_fullnoisemes}
\alias{divime_fullnoisemes}
\title{run the DiViMe noisemesFull SAD module}
\usage{
divime_fullnoisemes(audio_loc, divime_loc, vmstart = TRUE,
  vmshutdown = TRUE, messages = TRUE, overwrite = FALSE)
}
\arguments{
\item{audio_loc}{character, path to the audio files}

\item{divime_loc}{character, path to the DiViMe directory with a VM}

\item{vmstart}{logical, perform a check whether the VM is running and if not start it up (by default \code{TRUE}). Turning this off, will speed up the function a little bit, but requires that you are sure that the VM is indeed running in \code{divime_loc}.}

\item{vmshutdown}{logical, should the VM shut down after the operations are done (by default \code{TRUE})}

\item{messages}{logical, should the file names of each processed file be printed}

\item{overwrite}{logical, should output files be overwritten if they already exist (default is \code{FALSE})}
}
\value{
a data.frame with the locations of the created rttm files and some diagnostics
}
\description{
run the DiViMe noisemesFull SAD module
}
\details{
in contrast to \code{divime_sad()} this function returns all the different noiseme classes
}
\examples{
\dontrun{
audio_loc <- "~/Desktop/test_audio/smallfiles/"
divime_loc <- "/Volumes/Data/VM2/ooo/DiViMe"
res <- divime_fullnoisemes(audio_loc = audio_loc, divime_loc = divime_loc, vmshutdown = FALSE)
}

}
