% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divime_sad.R
\name{divime_sad}
\alias{divime_sad}
\title{run a DiViMe SAD module}
\usage{
divime_sad(audio_loc, divime_loc, module = "noisemes",
  splitaudio = FALSE, vmstart = TRUE, vmshutdown = TRUE,
  messages = TRUE, overwrite = FALSE)
}
\arguments{
\item{audio_loc}{character, path to the audio files}

\item{divime_loc}{character, path to the DiViMe directory with a VM}

\item{module}{character, which module to execute (default is \code{"noisemes"}), see details}

\item{splitaudio}{logical, should audio files be split into smaller chunks, default is \code{FALSE}, see details}

\item{vmstart}{logical, perform a check whether the VM is running and if not start it up (by default \code{TRUE}). Turning this off, will speed up the function a little bit, but requires that you are sure that the VM is indeed running in \code{divime_loc}.}

\item{vmshutdown}{logical, should the VM shut down after the operations are done (by default \code{TRUE})}

\item{messages}{logical, should the file names of each processed file be printed}

\item{overwrite}{logical, should output files be overwritten if they already exist (default is \code{FALSE})}
}
\value{
a data.frame with the locations of the created rttm files and some diagnostics
}
\description{
run a DiViMe SAD module
}
\details{
\code{module=} sets the SAD module to be used: can be either \code{"noisemes"}, \code{"opensmile"} or \code{"tocombo"}

It appears that some of the modules have difficulties with larger audio files (opensmile and noisemes). Hence, setting \code{split_audio} to \code{TRUE} will split the source audio into chunks of 2 minutes temporarilly. Note that this step requires the \code{sox} utility available (see \code{\link{set_binaries}} and \code{\link{split_audio}}).
}
