% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divime_sad_noisemes.R
\name{divime_sad_noisemes}
\alias{divime_sad_noisemes}
\title{run the DiBiMe SAD module noisemes}
\usage{
divime_sad_noisemes(audio_loc, divime_loc, cleanup = TRUE)
}
\arguments{
\item{audio_loc}{character, path to the audio files}

\item{divime_loc}{character, path to the DiViMe directory with running VM}

\item{cleanup}{logical, should the VM be shut down after the operations are done (by default \code{TRUE})}
}
\value{
a data.frame with the locations of the created rttm files
}
\description{
run the DiBiMe SAD module noisemes
}
