% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divime_test.R
\name{divime_test}
\alias{divime_test}
\title{test DiViMe VM}
\usage{
divime_test(divime_loc)
}
\arguments{
\item{divime_loc}{path where DiViMe is installed}
}
\value{
textual output
}
\description{
test DiViMe VM
}
