% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divime_vagrant_state.R
\name{divime_vagrant_state}
\alias{divime_vagrant_state}
\title{status and handling of VM}
\usage{
divime_vagrant_state(divime_loc, what = "status", silent = FALSE)
}
\arguments{
\item{divime_loc}{character, path to DiViMe directory}

\item{what}{character, what to do: see details}

\item{silent}{logical, should a message be printed}
}
\value{
a message (or modifies the state of the VM)
}
\description{
status and handling of VM
}
\details{
\code{what=} can be one of the following:
\itemize{
 \item \code{"status"}: default, what is the state of the VM

 \item \code{"start"}: start the VM

 \item \code{"suspend"}: suspend the VM

 \item \code{"halt"}: shut down the VM
}

in case of \code{what = "status"}, there is also an invisible result that indicates whether the VM is running
}
