% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divime_vagrant_update.R
\name{divime_vagrant_update}
\alias{divime_vagrant_update}
\title{update/re-install DiViMe VM}
\usage{
divime_vagrant_update(divime_loc, memoryoverride = 2048)
}
\arguments{
\item{divime_loc}{character, path to DiViMe directory}

\item{memoryoverride}{numeric value for the memory allocated to the VM (by default 2048, i.e. the default setting from DiViMi), see details}
}
\value{
creates a directory, clones the DiViMe repository into it and starts the virtual machine
}
\description{
update/re-install DiViMe VM
}
\details{
this function destroys (deletes) the virtual machine, pulls the git repository again and installs the virtual machine

with the default value allocated to the VM's memory, you might run into trouble analyzing larger files. Apparently a value of 4096 should work (although I personally use 6144 to be safe). Note that you probably shouldn't use more than about half of the memory your actual machine has...

Also note that this process can take a long time to execute (expect at least 20 min up to one hour depending on your internet connection and computer power)
}
