% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divime_wordcount.R
\name{divime_wordcount}
\alias{divime_wordcount}
\title{run the DiViMe word count estimator}
\usage{
divime_wordcount(audio_loc, divime_loc, speech_annos = "opensmile",
  vmstart = TRUE, vmshutdown = TRUE, messages = TRUE,
  overwrite = FALSE, ...)
}
\arguments{
\item{audio_loc}{character, path to the audio files}

\item{divime_loc}{character, path to the DiViMe directory with running VM}

\item{speech_annos}{character, what kind of speech detection is available, see details}

\item{vmstart}{logical, perform a check whether the VM is running and if not start it up (by default \code{TRUE}). Turning this off, will speed up the function a little bit, but requires that you are sure that the VM is indeed running in \code{divime_loc}.}

\item{vmshutdown}{logical, should the VM be shut down after the operations are done (by default \code{TRUE})}

\item{messages}{logical, should the file names of each processed file be printed}

\item{overwrite}{logical, should output files be overwritten if they already exist (default is \code{FALSE})}

\item{...}{further arguments for \code{\link{copy_audio}}, see details}
}
\value{
a data.frame with the locations of the created rttm files and some diagnostics
}
\description{
run the DiViMe word count estimator
}
\details{
\code{speech_annos} needs to be one of the following: \code{"noisemes"}, \code{"opensmile"}, \code{"tocombo"} or \code{"yunitator_english"}. The first three can be obtained from \code{\link{divime_sad}} and the latter from \code{\link{divime_talkertype}}.

In case you choose \code{speech_annos = "yunitator_english"} you might not get results unless you append your source audio file with some silence. This can be done inside the function here, and requires you setting two additional arguments to \code{divime_wordcount}: \code{appendsilence = 5} adds 5 seconds of silence, and \code{pathtosox=} requires the location of \code{sox}. See \code{\link{copy_audio}} and \code{\link{set_binaries}} for more details.
}
