% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elan2rttm.R
\name{elan2rttm}
\alias{elan2rttm}
\title{convert ELAN .eaf to .rttm}
\usage{
elan2rttm(x, outpath = NULL, clean_at = TRUE, use_py = TRUE)
}
\arguments{
\item{x}{character, path to elan (.eaf) file}

\item{outpath}{character, path to output directory (if \code{NULL} (the default), file is written in the same location as the input)}

\item{clean_at}{logical, remove tiers with @ in their name (dependent tiers)}

\item{use_py}{logical, use DiViMe python script (per default \code{TRUE}), or R style}
}
\value{
a file is written and the path to that file is returned as character
}
\description{
convert ELAN .eaf to .rttm
}
\details{
to install the required python package use \code{py_install("pympi-ling")}.

Also note that the \code{output=} argument needs to be a folder location. The output file name is 'fixed' in the sense that it's the same as the input, except that file extension is changed to .rttm.

If you use the python version of this function (\code{use_py = TRUE}), the ELAN files need to meet at least the following requirement: the tiers must not only have a 'tier name', but also must have a 'particpant' assigned to each tier. The underlying \code{pympi} python package will throw a message 'parsing unknown version of ELAN spec... This could result in errors ...', in case you use \emph{not} ELAN 2.6 or 2.7. Since current ELAN version are far above that, there is no way

The python version also re-sorts the output, i.e. entries are sorted according to tier name whereas the R version sorts according to starting time.

If a file exists with the same name as the target file, it will be overwritten.
}
\examples{
\dontrun{
tdir <- tempdir()
elanfile <- system.file("spanish.eaf", package = "avutils")

# use R
dest <- elan2rttm(x = elanfile, outpath = tdir, use_py = FALSE)
"spanish.rttm" \%in\% list.files(tdir)
temp <- read_rttm(paste0(tdir, "/spanish.rttm"))
temp <- temp[order(temp$start), ]
head(temp)
file.remove(paste0(tdir, "/spanish.rttm"))

# use python internally
dest <- elan2rttm(x = elanfile, outpath = tdir, use_py = TRUE)
"spanish.rttm" \%in\% list.files(tdir)
temp <- read_rttm(paste0(tdir, "/spanish.rttm"))
temp <- temp[order(temp$start), ]
head(temp)
file.remove(paste0(tdir, "/spanish.rttm"))
}
}
