% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prefix.R
\name{get_prefix}
\alias{get_prefix}
\title{get DiViMe prefixes for all rttm files in a location}
\usage{
get_prefix(loc, include_subfolders = FALSE)
}
\arguments{
\item{loc}{character, path that contains .rttm files}

\item{include_subfolders}{logical, include files in subfolder (default is \code{FALSE})}
}
\value{
a data frame
}
\description{
get DiViMe prefixes for all rttm files in a location
}
\examples{
X <- dirname(list.files(system.file("", package = "avutils"), full.names = TRUE)[1])
get_prefix(X)
# with files in subfolders:
tdir <- tempdir()
dir.create(file.path(tdir, "test"))
dir.create(file.path(tdir, "test/subfolder"))
file.copy(from = system.file("yunitator_old_spanish.rttm", package = "avutils"),
          to = file.path(tdir, "test/subfolder/yunitator_old_spanish.rttm"))
file.copy(from = system.file("tocomboSad_BER_0485_12_07_09123.rttm", package = "avutils"),
          to = file.path(tdir, "test/tocomboSad_BER_0485_12_07_09123.rttm"))
get_prefix(file.path(tdir, "test"))
get_prefix(file.path(tdir, "test"), include_subfolders = TRUE)
}
