% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_filenames.R
\name{handle_filenames}
\alias{handle_filenames}
\title{file and folder names}
\usage{
handle_filenames(audio_loc, divime_loc)
}
\arguments{
\item{audio_loc}{character, path to the audio files}

\item{divime_loc}{character, path to the DiViMe directory with a VM}
}
\value{
a data.frame with various versions of file paths and file names
}
\description{
file and folder names
}
\details{
the functions also cleans file names (temporarily) to get rid of spaces and special characters in file names, which otherwise would cause problems with some of the DiViMe tools
}
