% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadingzeros.R
\name{leadingzeros}
\alias{leadingzeros}
\title{add zeros to number as string for composing file names}
\usage{
leadingzeros(num, ndigits, ext = NULL)
}
\arguments{
\item{num}{numeric or character, the number}

\item{ndigits}{numeric, the total number of digits for the output}

\item{ext}{character, optional, a file extension to be added}
}
\value{
character
}
\description{
add zeros to number as string for composing file names
}
\examples{
leadingzeros(3, 2)
leadingzeros(3, 5)
leadingzeros(333, 3)
leadingzeros(333, 6)
leadingzeros(333, 6, ".wav")
# produces an error
# leadingzeros(333, 2)

# similar to
sprintf("\%06.0f", 3)
}
