% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_sads.R
\name{pair_sads}
\alias{pair_sads}
\title{visual comparison of speech analysers}
\usage{
pair_sads(x1, x2, nlines = 5, from = NULL, to = NULL,
  overlap = NULL, ...)
}
\arguments{
\item{x1}{character, path to file with annotations in either ELAN (.eaf) or DiViMe (.rttm) format}

\item{x2}{character, path to file with annotations in either ELAN (.eaf) or DiViMe (.rttm) format}

\item{nlines}{numeric, the number of lines in output graphic}

\item{from}{numeric, starting point for selecting a subset of time}

\item{to}{numeric, starting point for selecting a subset of time}

\item{overlap}{numeric, the time increment for calculating overlap, by default \code{NULL}, i.e. no overlap is calculated}

\item{...}{additional parameters for elan files (which tiers/annotations to ignore), see \code{\link{collapse_tiers}}}
}
\value{
a numeric vector with proportions of agreement between the two SADs
}
\description{
visual comparison of speech analysers
}
