% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praat_pitch.R
\name{praat_pitch}
\alias{praat_pitch}
\title{extract some pitch variables via Praat from audio (segments)}
\usage{
praat_pitch(sound_file, child = TRUE, time_range = NULL,
  pathtopraat = NULL, unit = "Hertz", pathtosox = NULL)
}
\arguments{
\item{sound_file}{character, path to audio file}

\item{child}{logical, should pitch settings be set for child (default),
or adult}

\item{time_range}{numeric of length 2, optional start and end time in
seconds. If left blank (\code{NULL}) the entire file is processed.}

\item{pathtopraat}{character, the path to the Praat binary}

\item{unit}{character, the unit for the frequency measures (\code{"Hertz"},
\code{"ERB"}, ...)}

\item{pathtosox}{character, optional path to sox binary (only required if
\code{time_range} is \emph{not} \code{NULL})}
}
\value{
a data frame with one row
}
\description{
extract some pitch variables via Praat from audio (segments)
}
\details{
The output reflect measure of F0: mean, 5% percentile and 95%
percentile.

The pitch settings for child versus adult are the following. Child:
floor = 150 Hz, ceiling = 650 Hz. Adult: floor = 75 Hz, ceiling = 400 Hz.
Time step for both is 0.01 s.

For this function to work, you need Praat, and you need to know where the
\code{Praat} binary is located. On a Mac the location is likely to be
"/Applications/Praat.app/Contents/MacOS/Praat", and on Windows
"C:/Program Files/Praat.exe".
}
\examples{
\dontrun{
# requires Praat installed
# location to Praat on my machine:
ploc <- "/Applications/Praat.app/Contents/MacOS/Praat"
# example sound
sound_file <- system.file("synthetic_speech.wav", package = "avutils")
praat_pitch(sound_file, child = FALSE, pathtopraat = ploc)
praat_pitch(sound_file, child = TRUE, pathtopraat = ploc)
# analyse only subset requires sox installed
sloc <- "/usr/local/bin/sox"
praat_pitch(sound_file, time_range = c(0.3, 3.1), child = FALSE,
            pathtopraat = ploc, pathtosox = sloc)
}
}
