% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_elan.R
\name{read_elan}
\alias{read_elan}
\title{read annotations from Elan .eaf file}
\usage{
read_elan(x)
}
\arguments{
\item{x}{either character string, providing the path to the elan file, or an \code{xml_document} (read by \code{xml2::read_xml()})}
}
\value{
a data frame
}
\description{
read annotations from Elan .eaf file
}
