% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rttm_pairs.R
\name{rttm_pairs}
\alias{rttm_pairs}
\title{find pairs of rttm files suitable for evaluations}
\usage{
rttm_pairs(test = NULL, reference = NULL, location = NULL,
  prefix = NULL, include_subfolders = FALSE)
}
\arguments{
\item{test}{character, file paths to reference files, i.e. rttm fiels with prefix}

\item{reference}{character, file paths to test files, i.e. rttm files without prefix}

\item{location}{character, path that contains .rttm files}

\item{prefix}{character, optionally, select only pairs for these prefixes. Default is \code{NULL}, i.e. all prefixes will be returned}

\item{include_subfolders}{logical, include files in subfolder (default is \code{FALSE})}
}
\value{
a data.frame with pairs of files
}
\description{
find pairs of rttm files suitable for evaluations
}
\details{
The function works in one of two ways. First, you can specify a \code{location=} in which all rttm files will be checked.

Second, you supply vectors for \code{test=} and \code{reference=} files.

The way via \code{location=} takes precedence, i.e. if you supply all three arguments, only the \code{location=} option is used.
}
\examples{
# using location
location <- dirname(list.files(system.file("", package = "avutils"), full.names = TRUE)[1])
rttm_pairs(location = location)
rttm_pairs(location = location, prefix = c("tocomboSad_"))

# using test and reference
reference <- system.file("BER_0485_12_07_09123.rttm", package = "avutils")
test <- list.files(system.file("", package = "avutils"), pattern = ".rttm$", full.names = TRUE)
rttm_pairs(test = test, reference = reference)
rttm_pairs(test = test, reference = reference, prefix = "tocomboSad_")
}
