% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_annotations.R
\name{segment_annotations}
\alias{segment_annotations}
\title{segment annotations time-based intervals}
\usage{
segment_annotations(xdata, segment_dur, timecols = c("start", "end"))
}
\arguments{
\item{xdata}{data.frame, the data object with annotations (e.g. import from ELAN or .rttm file)}

\item{segment_dur}{numeric, the time interval in the relevant unit (no default, see details)}

\item{timecols}{character, the column names for start and end of the annotations (default is \code{c("start", "end")}).}
}
\value{
a data.frame with a new column \code{cat} that indicates which time interval a given annotation belongs to
}
\description{
assign each annotation to regular (e.g. one minute) segments
}
\details{
note that annotations that cross over a given cut point for a segment will be split into two (or more if an annotation is much longer than the segment duration)
}
\examples{
annos <- LETTERS[1:5]
start <- c(14, 17, 45, 65, 70)
end <- c(25, 22, 60, 80, 73)
dur <- end - start
xdata <- data.frame(start, end, dur, annos)

xdata
segment_annotations(xdata, segment_dur = 10, timecols = c("start", "end"))

# from an ELAN file:
eaf <- system.file("synthetic_speech.eaf", package = "avutils")
elanfile <- read_elan(eaf)
segment_annotations(elanfile, segment_dur = 0.5)
}
