% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_divime.R
\name{setup_divime}
\alias{setup_divime}
\title{setup DiViMe}
\usage{
setup_divime(divime_loc)
}
\arguments{
\item{divime_loc}{path where the virtual machine should be installed, directory will be created if it doesn't exist}
}
\value{
creates a directory, clones the DiViMe repository into it and starts the virtual machine
}
\description{
setup DiViMe
}
