% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_audio.R
\name{split_audio}
\alias{split_audio}
\title{split audio file into chunks of defined length}
\usage{
split_audio(filein, split, pathout = NULL,
  pathtosox = getOption("avutils_sox"))
}
\arguments{
\item{filein}{character, path to input audio}

\item{split}{numeric, duration of split in seconds}

\item{pathout}{character, path were output is stored, by default the same as the source}

\item{pathtosox}{character, path to sox binary}
}
\value{
a vector with the paths to the split files
}
\description{
split audio file into chunks of defined length
}
