% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_duration.R
\name{split_duration}
\alias{split_duration}
\title{split annotations into time-based intervals (e.g. per minute)}
\usage{
split_duration(xdata, splitduration = 60000, timecols = c("start",
  "end"), end_is_dur = FALSE)
}
\arguments{
\item{xdata}{data.frame, the data object with annotations (e.g. import from ELAN or .rttm file)}

\item{splitduration}{numeric, the time interval in milliseconds (default is 60000, i.e. one minute)}

\item{timecols}{character, the column names for start and end of the annotations}

\item{end_is_dur}{logical, is the column with the end time actually a duration (by default \code{FALSE}, i.e. end column is treated as a time stamp)}
}
\value{
a data.frame with a new column that indicates which time interval a given annotation belongs to
}
\description{
split annotations into time-based intervals (e.g. per minute)
}
\details{
note that annotations that cross over a given cut point will be split into two

if \code{end_is_dur = TRUE} a new column ('\code{xend}') is added that adds the second time column (which in that case should be a duration) to the start time
}
\examples{
annos <- LETTERS[1:5]
start <- c(14, 17, 45, 65, 70)
end <- c(25, 22, 63, 80, 73)

xdata <- data.frame(start, end, annos)

xdata
split_duration(xdata, splitduration = 30, timecols = c("start", "end"))

# from an ELAN file:
eaf <- system.file("synthetic_speech.eaf", package = "avutils")
elanfile <- read_elan(eaf)
split_duration(elanfile, splitduration = 5000)
}
