% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_time.R
\name{subset_time}
\alias{subset_time}
\title{select subset of elan or rttm annotations}
\usage{
subset_time(x, from = NULL, to = NULL, include_incomplete = TRUE,
  shift_times = FALSE)
}
\arguments{
\item{x}{a data.frame (result from \code{\link{read_elan}} or \code{\link{read_rttm}})}

\item{from}{numeric, include annotations from this point forward (by default \code{NULL}, i.e. take everything from the first line of the data)}

\item{to}{numeric, include annotations up to that point (by default \code{NULL}, i.e. take up to the last line)}

\item{include_incomplete}{logical, should annotations that cross \code{from = } and/or \code{to = } be kept (by default \code{TRUE})}

\item{shift_times}{logical, not implemented (yet)}
}
\value{
a data.frame
}
\description{
select subset of elan or rttm annotations
}
\details{
If you keep annotations that cross the from and/or to boundaries, the durations of those annotations will be updated.
}
\examples{
rttm <- system.file("yunitator_old_synthetic_speech_overlap.rttm", package = "avutils")
x <- read_rttm(rttm)
# leave unchanged
subset_time(x)
# up to 26
subset_time(x, to = 26)
# remove incomplete anno (25.6 - 26.3)
subset_time(x, to = 26, include_incomplete = FALSE)

elan <- system.file("synthetic_speech_overlap.eaf", package = "avutils")
x <- read_elan(elan)
# leave unchanged
subset_time(x)
# up to 29
subset_time(x, to = 29)
# remove incomplete anno (28.3 - 30.5)
subset_time(x, to = 29, include_incomplete = FALSE)
}
