% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_elan.R
\name{summary_elan}
\alias{summary_elan}
\title{summarize Elan .eaf file}
\usage{
summary_elan(x)
}
\arguments{
\item{x}{either character string, providing the path to the elan file(s)}
}
\value{
a list with two elements. The first is a data.frame with summaries and the second is a character vector with annotations ordered by their number of occurences (first is the most frequent)
}
\description{
summarize Elan .eaf file
}
