% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_binaries.R
\name{test_binaries}
\alias{test_binaries}
\title{test whether sox, ffmpeg, git and vagrant binaries work}
\usage{
test_binaries(printmessages = TRUE)
}
\arguments{
\item{printmessages}{logical, should messages be returned}
}
\value{
a message
}
\description{
test whether sox, ffmpeg, git and vagrant binaries work
}
