% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_divime.R
\name{test_divime}
\alias{test_divime}
\title{test DiViMe VM}
\usage{
test_divime(divime_loc)
}
\arguments{
\item{divime_loc}{path where DiViMe installed}
}
\value{
textual output
}
\description{
test DiViMe VM
}
