% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_info.R
\name{video_info}
\alias{video_info}
\title{get media information}
\usage{
video_info(filein, pathtoffmpeg = getOption("avutils_ffmpeg"))
}
\arguments{
\item{filein}{character, path to audio file}

\item{pathtoffmpeg}{character, path to ffmpeg binary}
}
\value{
a data.frame with features of video files
}
\description{
video information via ffmpeg
}
