% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add2elan.R
\name{add2elan}
\alias{add2elan}
\title{add DiViMe SAD annotations to ELAN file}
\usage{
add2elan(elanfile, audiofile, tocombo = NULL, noisemes = NULL,
  opensmile = NULL, noisemesFull = NULL)
}
\arguments{
\item{elanfile}{character, path to elan file}

\item{audiofile}{character, path to audio file}

\item{tocombo}{character, path to tocombo rttm (optional)}

\item{noisemes}{character, path to noisemes rttm (optional)}

\item{opensmile}{character, path to opensmile rttm (optional)}

\item{noisemesFull}{character, path to noisemesFull rttm (optional)}
}
\value{
a file is created in the same location as \code{audiofile} and the path to it is returned as character
}
\description{
add DiViMe SAD annotations to ELAN file
}
\details{
This is bascially a wrapper for \code{\link{rttm2elan}}, which takes care of homogenizing the format of .rttm files when they come from different sources (ELAN and DiViMe tools).

Note also that this function can take a long time to finish (due to the conversion from an intermediate .rttm file to ELAN .eaf).
}
