% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_info.R
\name{audio_info}
\alias{audio_info}
\title{get information about audio file via sox}
\usage{
audio_info(filein, pathtosox = getOption("avutils_sox"))
}
\arguments{
\item{filein}{character, path to audio file}

\item{pathtosox}{character, path to sox binary}
}
\value{
a list with features of audio file
}
\description{
get information about audio file via sox
}
