% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_info.R
\name{audio_info}
\alias{audio_info}
\title{get media information}
\usage{
audio_info(filein, pathtosox = getOption("avutils_sox"))
}
\arguments{
\item{filein}{character, path to audio files}

\item{pathtosox}{character, path to sox binary}
}
\value{
a dataframe with features of audio files
}
\description{
audio information via sox
}
