% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_audio.R
\name{convert_audio}
\alias{convert_audio}
\title{convert audio file format}
\usage{
convert_audio(filein, outformat = list(filetype = "wav", res = 44100,
  mono = TRUE), overwrite = FALSE, pathout = NULL,
  pathtoffmpeg = getOption("avutils_ffmpeg"))
}
\arguments{
\item{filein}{character, path to input audio}

\item{outformat}{a list with specifications of the desired output format, see details}

\item{overwrite}{logical, should output files be overwritten if it already exists}

\item{pathout}{character, path were output is stored, by default the same as the source}

\item{pathtoffmpeg}{character, path to ffmpeg binary}
}
\value{
a vector with the paths to the split files
}
\description{
convert audio file format
}
\details{
the \code{outformat=} argument currently provides four options, and by default a mono (\code{mono = TRUE}) .wav files is produced (\code{filetpye = "wav"}) with 44.1kHz resolution (\code{res = 441000}).
}
