% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divime_classify_vox.R
\name{divime_classify_vox}
\alias{divime_classify_vox}
\title{run the DiViMe vocal classification module vcm}
\usage{
divime_classify_vox(audio_loc, divime_loc, vmshutdown = TRUE,
  messages = TRUE, overwrite = FALSE)
}
\arguments{
\item{audio_loc}{character, path to the audio files}

\item{divime_loc}{character, path to the DiViMe directory with running VM}

\item{vmshutdown}{logical, should the VM be shut down after the operations are done (by default \code{TRUE})}

\item{messages}{logical, should the file names of each processed file be printed}

\item{overwrite}{logical, should output files be overwritten if they already exist (default is \code{FALSE})}
}
\value{
a data.frame with the locations of the created rttm files and some diagnostics
}
\description{
run the DiViMe vocal classification module vcm
}
