% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divime_run_all.R
\name{divime_run_all}
\alias{divime_run_all}
\title{run all DiViMe modules}
\usage{
divime_run_all(audio_loc, divime_loc, vmstart = TRUE,
  vmshutdown = TRUE, messages = TRUE, overwrite = FALSE)
}
\arguments{
\item{audio_loc}{character, path to the audio files}

\item{divime_loc}{character, path to the DiViMe directory with a VM}

\item{vmstart}{logical, perform a check whether the VM is running and if not start it up (by default \code{TRUE}). Turning this off, will speed up the function a little bit, but requires that you are sure that the VM is indeed running in \code{divime_loc}.}

\item{vmshutdown}{logical, should the VM shut down after the operations are done (by default \code{TRUE})}

\item{messages}{logical, should the file names of each processed file be printed}

\item{overwrite}{logical, should output files be overwritten if they already exist (default is \code{FALSE})}
}
\value{
a list of data.frames with the locations of the created rttm files and some diagnostics
}
\description{
run all DiViMe modules
}
