% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_roles.R
\name{evaluate_roles}
\alias{evaluate_roles}
\title{evaluate speaker roles}
\usage{
evaluate_roles(xdata, resolution = 1, duration = NULL, tiers = NULL)
}
\arguments{
\item{xdata}{a data.frame, (result from \code{\link{read_elan}} or \code{\link{read_rttm}})}

\item{resolution}{numeric, the time increment in seconds, by default 1}

\item{duration}{numeric, optional info about the duration of the audio. If not supplied, the end of the last annotation is taken as the duration}

\item{tiers}{character of variable length, the tiers that should be considered}
}
\value{
a matrix
}
\description{
evaluate speaker roles
}
\examples{
elan <- system.file("synthetic_speech_overlap.eaf", package = "avutils")
xdata <- read_elan(elan)
evaluate_roles(xdata)
}
