% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_sad.R
\name{evaluate_sad}
\alias{evaluate_sad}
\title{evaluation of speech detection tools}
\usage{
evaluate_sad(x1, x2, nsegments = 1, from = NULL, to = NULL,
  overlap_res = NULL, doplot = TRUE, raw_results = FALSE, ...)
}
\arguments{
\item{x1}{character, path to file with annotations in either ELAN (.eaf) or DiViMe (.rttm) format}

\item{x2}{character, path to file with annotations in either ELAN (.eaf) or DiViMe (.rttm) format}

\item{nsegments}{numeric, the number of segments in which the file is to be split (represented as different lines in output graphic). Default is \code{1}, i.e. the entire file is considered as one segment.}

\item{from}{numeric, starting point for selecting a subset of time}

\item{to}{numeric, starting point for selecting a subset of time}

\item{overlap_res}{numeric, the time resolution for calculating overlap, by default \code{NULL}, i.e. no overlap is calculated, see \code{\link{annotation_overlap}}}

\item{doplot}{logical, should the plot actually be produced (default is \code{TRUE})}

\item{raw_results}{logical, should the raw results be returnd, i.e. the values per 'sample' (by default \code{FALSE})}

\item{...}{additional parameters for elan files (which tiers/annotations to ignore), see \code{\link{collapse_tiers}}}
}
\value{
a data.frame with proportions of agreement between the two SADs
}
\description{
evaluation of speech detection tools
}
\examples{
x1 <- system.file("noisemesSad_synthetic_speech.rttm", package = "avutils")
x2 <- system.file("synthetic_speech.eaf", package = "avutils")
# no overlap calculation
evaluate_sad(x1 = x1, x2 = x2)
# fairly crude time resolution for overlap calculation
evaluate_sad(x1 = x1, x2 = x2, overlap_res = 10)
# slightly higher time resolution
evaluate_sad(x1 = x1, x2 = x2, overlap_res = 100)
# even more fine-grained
evaluate_sad(x1 = x1, x2 = x2, overlap_res = 1000)
# even more fine-grained and file split into multiple segments
evaluate_sad(x1 = x1, x2 = x2, nsegments = 10, overlap_res = 100)
}
