% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_audio.R
\name{extract_audio}
\alias{extract_audio}
\title{extract mono wav file from video source}
\usage{
extract_audio(videofile, pathout = NULL,
  pathtoffmpeg = getOption("avutils_ffmpeg"))
}
\arguments{
\item{videofile}{character, path to video file}

\item{pathout}{character, path where the audio should be saved (by default the same as the video file)}

\item{pathtoffmpeg}{character, the path to the ffmpeg binary}
}
\value{
a character string with the path to the written audio file
}
\description{
extract mono wav file from video source
}
