% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_sads.R
\name{pair_sads}
\alias{pair_sads}
\title{visual comparison of speech analysers}
\usage{
pair_sads(x1, x2, nlines = 5, from = NULL, to = NULL,
  overlap_res = NULL, ...)
}
\arguments{
\item{x1}{character, path to file with annotations in either ELAN (.eaf) or DiViMe (.rttm) format}

\item{x2}{character, path to file with annotations in either ELAN (.eaf) or DiViMe (.rttm) format}

\item{nlines}{numeric, the number of lines in output graphic}

\item{from}{numeric, starting point for selecting a subset of time}

\item{to}{numeric, starting point for selecting a subset of time}

\item{overlap_res}{numeric, the time resolution for calculating overlap, by default \code{NULL}, i.e. no overlap is calculated, see \code{\link{annotation_overlap}}}

\item{...}{additional parameters for elan files (which tiers/annotations to ignore), see \code{\link{collapse_tiers}}}
}
\value{
a numeric vector with proportions of agreement between the two SADs
}
\description{
visual comparison of speech analysers
}
\examples{
x1 <- system.file("synthetic_speech.eaf", package = "avutils")
x2 <- system.file("noisemesSad_synthetic_speech.rttm", package = "avutils")
# no overlap calculation
pair_sads(x1 = x1, x2 = x2, nlines = 2)
# fairly crude time resolution for overlap calculation
pair_sads(x1 = x1, x2 = x2, nlines = 2, overlap_res = 10)
# slightly higher time resolution
pair_sads(x1 = x1, x2 = x2, nlines = 2, overlap_res = 100)
# even more fine-grained
pair_sads(x1 = x1, x2 = x2, nlines = 2, overlap_res = 1000)
}
