% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rttm.R
\name{read_rttm}
\alias{read_rttm}
\title{read rttm files}
\usage{
read_rttm(x)
}
\arguments{
\item{x}{character, path to rttm file}
}
\value{
a data.frame
}
\description{
read and prettify rttm files
}
\details{
The function adds an end time column and assigns meaningful column names for the most relevant columns. The actual content of the eighth column (\code{tier}) depends on the rttm file. For example, 'yunitator' rttm files have levels \code{CHI}, \code{FEM} and \code{MAL}, while SAD outputs contain \code{speech} (or in the case of the full noisemes module other labels like \code{background}, \code{noise} etc).
}
\examples{
rttm <- system.file("noisemesSad_synthetic_speech_overlap.rttm", package = "avutils")
read_rttm(rttm)
rttm <- system.file("noisemesFull_synthetic_speech_overlap.rttm", package = "avutils")
read_rttm(rttm)
rttm <- system.file("yunitator_english_synthetic_speech_overlap.rttm", package = "avutils")
read_rttm(rttm)
rttm <- system.file("tocomboSad_synthetic_speech_overlap.rttm", package = "avutils")
read_rttm(rttm)
}
