% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_annotations.R
\name{segment_annotations}
\alias{segment_annotations}
\title{segment annotations time-based intervals}
\usage{
segment_annotations(xdata, segment_dur, timecols = c("start", "end"),
  end_is_dur = FALSE)
}
\arguments{
\item{xdata}{data.frame, the data object with annotations (e.g. import from ELAN or .rttm file)}

\item{segment_dur}{numeric, the time interval in the relevant unit (no default, see details)}

\item{timecols}{character, the column names for start and end of the annotations (default is \code{c("start", "end")}). If there is no 'end' time stamp but only a duration for the annotation, use this column.}

\item{end_is_dur}{logical, is the column with the end time actually a duration (by default \code{FALSE}, i.e. end column is treated as a time stamp)}
}
\value{
a data.frame with a new column that indicates which time interval a given annotation belongs to
}
\description{
assign each annotation to regular (e.g. one minute) segments
}
\details{
the unit of the \code{segment_dur=} argument depends on the unit of time in the source file. ELAN uses milliseconds, while the DiViMe tools typically use seconds.

note that annotations that cross over a given cut point for a segment will be split into two (or more if an annotation is much longer than the segment duration)

if \code{end_is_dur = TRUE} a new column ('\code{xend}') is added that adds the second time column (which in that case should be a duration) to the start time
}
\examples{
annos <- LETTERS[1:5]
start <- c(14, 17, 45, 65, 70)
end <- c(25, 22, 60, 80, 73)
dur <- end - start
xdata <- data.frame(start, end, dur, annos)

xdata
segment_annotations(xdata, segment_dur = 10, timecols = c("start", "end"))

# from an ELAN file:
eaf <- system.file("synthetic_speech.eaf", package = "avutils")
elanfile <- read_elan(eaf)
segment_annotations(elanfile, segment_dur = 500)
}
