% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_duration.R
\name{split_duration}
\alias{split_duration}
\title{split annotations into time-based intervals (e.g. per minute)}
\usage{
split_duration(xdata, splitduration = 60000, timecols = c("start",
  "end"))
}
\arguments{
\item{xdata}{data.frame, the data object with annotations (e.g. import from ELAN or .rttm file)}

\item{splitduration}{numeric, the time interval in milliseconds (default is 60000, i.e. one minute)}

\item{timecols}{character, the column names for start and end of the annotations}
}
\value{
a data.frame with a new column that indicates which time interval a given annotation belongs to
}
\description{
split annotations into time-based intervals (e.g. per minute)
}
\details{
note that annotations that cross over a given cut point will be split into two
}
