% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_elan.R
\name{summary_elan}
\alias{summary_elan}
\title{summarize Elan .eaf file}
\usage{
summary_elan(x)
}
\arguments{
\item{x}{either character string, providing the path to the elan file, or an \code{xml_document} (read by \code{xml2::read_xml()})}
}
\value{
a data frame
}
\description{
summarize Elan .eaf file
}
